fix:: 채팅 페이지 CSR 범위 최소화
